+++ src/cmd/compile/internal/ssa/_gen/ARM64.rules
@@ -1931,7 +1931,7 @@
 
 ((REV|REVW) ((REV|REVW) p)) => p
 
-// runtime/internal/math.MulUintptr intrinsics
+// internal/runtime/math.MulUintptr intrinsics
 
 (Select0 (Mul64uover x y)) => (MUL x y)
 (Select1 (Mul64uover x y)) => (NotEqual (CMPconst (UMULH <typ.UInt64> x y) [0]))
