+++ src/runtime/unsafe.go
@@ -5,7 +5,7 @@
 package runtime
 
 import (
-	"runtime/internal/math"
+	"internal/runtime/math"
 	"unsafe"
 )
